requirements.txt
absl-py==0.11.0
aiohttp==3.7.2
aiohttp-cors==0.7.0
appnope==0.1.0
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.2.0
backcall==0.2.0
cached-property==1.5.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorclass==2.2.0
colorlog==4.4.0
croniter==0.3.35
decorator==4.4.2
envyaml==0.2060
faust==1.10.4
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
idna==2.10
importlib-metadata==2.0.0
injectable==3.0.1
ipdb==0.13.4
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
joblib==0.17.0
jsonschema==3.2.0
kafka-python==1.4.7
Keras==2.4.3
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
Markdown==3.3.3
mock==4.0.2
mode==4.3.2
multidict==5.0.0
mypy-extensions==0.4.3
natsort==7.0.1
nltk==3.5
numpy==1.18.5
oauthlib==3.1.0
opentracing==1.3.0
opt-einsum==3.3.0
parameters-validation==1.2.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycollect==0.2.2
Pygments==2.7.2
pyrsistent==0.17.3
python-chain==1.0.4
python-dateutil==2.8.1
PyYAML==5.3.1
regex==2020.10.28
requests==2.24.0
requests-oauthlib==1.3.0
robinhood-aiokafka==1.1.6
rsa==4.6
scipy==1.5.3
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminaltables==3.1.0
tqdm==4.51.0
traitlets==5.0.5
typing-extensions==3.7.4.3
typing-inspect==0.5.0
urllib3==1.25.11
venusian==1.2.0
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.2
zipp==3.4.0
